Module re_protos::codec

Codec for serializing and deserializing query response (record batch) data

Structs

MessageHader

Enums

CodecError

TransportMessageV0

Functions

decode

Decode transport data from a byte stream - if there’s a record batch present, return it, otherwise return None.

encode

Encode a transport chunk into a byte stream.

no_data

Encode a NoData message into a byte stream. This can be used by the remote store (i.e. data producer) to signal back to the client that there’s no data available.

read_arrow_from_bytes 🔒

Helper function that deserializes raw bytes into arrow schema and record batch using Arrow IPC format.

write_arrow_to_bytes 🔒

Helper function that serializes given arrow schema and record batch into bytes using Arrow IPC format.