re_protos::manifest_registry
Module v1alpha1

Modules

ext

index_properties

Nested message and enum types in IndexProperties.

index_query_properties

Nested message and enum types in IndexQueryProperties.

manifest_registry_service_client

Generated client implementations.

manifest_registry_service_server

Generated server implementations.

Structs

BTreeIndex

TODO(zehiko) add properties as needed

BTreeIndexQuery

TODO(zehiko) add properties as needed

CreateIndexRequest

CreateIndexResponse

CreatePartitionManifestsRequest

CreatePartitionManifestsResponse

DataSource

Error

Application level error - used as details in the google.rpc.Status message

FetchChunkManifestRequest

FetchChunkManifestResponse

FetchPartitionManifestRequest

FetchPartitionManifestResponse

TODO(cmc): this should have response extensions too.

GetChunksRequest

GetChunksResponse

GetDatasetSchemaRequest

GetDatasetSchemaResponse

GetPartitionTableSchemaRequest

GetPartitionTableSchemaResponse

IndexColumn

used to define which column we want to index

IndexConfig

IndexProperties

IndexQueryProperties

InvertedIndex

InvertedIndexQuery

TODO(zehiko) add properties as needed

Query

QueryDatasetRequest

QueryDatasetResponse

QueryLatestAt

A chunk-level latest-at query, aka LatestAtRelevantChunks.

QueryRange

/ A chunk-level range query, aka RangeRelevantChunks.

ReIndexRequest

ReIndexResponse

RegisterWithDatasetBlockingRequest

TODO(andrea): This is a copy of RegisterWithDatasetRequest. Eventually we may get rid of the sync version; until then, we should make sure that the two objects are in sync.

RegisterWithDatasetBlockingResponse

RegisterWithDatasetRequest

RegisterWithDatasetResponse

ScanPartitionTableRequest

ScanPartitionTableResponse

SearchDatasetRequest

SearchDatasetResponse

VectorIndexQuery

VectorIvfPqIndex

WriteChunksRequest

WriteChunksResponse

Enums

DataSourceKind

ErrorCode

Error codes for application level errors

VectorDistanceMetric

Constants

DATASET_MANIFEST_ID_FIELD_NAME

DatasetManifest mandatory field names. All mandatory metadata fields are prefixed with “rerun_” to avoid conflicts with user-defined fields.

DATASET_MANIFEST_PARTITION_MANIFEST_UPDATED_AT_FIELD_NAME

DATASET_MANIFEST_PARTITION_MANIFEST_URL_FIELD_NAME

DATASET_MANIFEST_RECORDING_TYPE_FIELD_NAME

DATASET_MANIFEST_REGISTRATION_TIME_FIELD_NAME

DATASET_MANIFEST_START_TIME_FIELD_NAME

DATASET_MANIFEST_STORAGE_URL_FIELD_NAME