Trait re_protos::remote_store::v0::storage_node_server::StorageNode

pub trait StorageNode: Send + Sync + 'static {
    type QueryStream: Stream<Item = Result<DataframePart, Status>> + Send + 'static;
    type FetchRecordingStream: Stream<Item = Result<RerunChunk, Status>> + Send + 'static;
    type SearchIndexStream: Stream<Item = Result<DataframePart, Status>> + Send + 'static;
    type QueryCatalogStream: Stream<Item = Result<DataframePart, Status>> + Send + 'static;

    // Required methods
    fn query<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::QueryStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn fetch_recording<'life0, 'async_trait>(
        &'life0 self,
        request: Request<FetchRecordingRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::FetchRecordingStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn create_index<'life0, 'async_trait>(
        &'life0 self,
        request: Request<CreateIndexRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CreateIndexResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn search_index<'life0, 'async_trait>(
        &'life0 self,
        request: Request<SearchIndexRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::SearchIndexStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn query_catalog<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryCatalogRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::QueryCatalogStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn update_catalog<'life0, 'async_trait>(
        &'life0 self,
        request: Request<UpdateCatalogRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<UpdateCatalogResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_recording_schema<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetRecordingSchemaRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetRecordingSchemaResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn register_recording<'life0, 'async_trait>(
        &'life0 self,
        request: Request<RegisterRecordingRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<DataframePart>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn unregister_recording<'life0, 'async_trait>(
        &'life0 self,
        request: Request<UnregisterRecordingRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<UnregisterRecordingResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn unregister_all_recordings<'life0, 'async_trait>(
        &'life0 self,
        request: Request<UnregisterAllRecordingsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<UnregisterAllRecordingsResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with StorageNodeServer.

Required Associated Types

type QueryStream: Stream<Item = Result<DataframePart, Status>> + Send + 'static

Server streaming response type for the Query method.

type FetchRecordingStream: Stream<Item = Result<RerunChunk, Status>> + Send + 'static

Server streaming response type for the FetchRecording method.

type SearchIndexStream: Stream<Item = Result<DataframePart, Status>> + Send + 'static

Server streaming response type for the SearchIndex method.

type QueryCatalogStream: Stream<Item = Result<DataframePart, Status>> + Send + 'static

Server streaming response type for the QueryCatalog method.

Required Methods

fn query<'life0, 'async_trait>( &'life0 self, request: Request<QueryRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::QueryStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

data API calls

fn fetch_recording<'life0, 'async_trait>( &'life0 self, request: Request<FetchRecordingRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::FetchRecordingStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn create_index<'life0, 'async_trait>( &'life0 self, request: Request<CreateIndexRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<CreateIndexResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn search_index<'life0, 'async_trait>( &'life0 self, request: Request<SearchIndexRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::SearchIndexStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

The response to SearchIndex a RecordBatch with 3 columns:

‘resource_id’ column with the id of the resource

timepoint column with the values representing the points in time where index query matches. What time points are matched depends on the type of index that is queried. For example for vector search it might be timepoints where top-K matches are found within each resource in the indexed entry. For inverted index it might be timepoints where the query string is found in the indexed column

‘data’ column with the data that is returned for the matched timepoints

fn query_catalog<'life0, 'async_trait>( &'life0 self, request: Request<QueryCatalogRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::QueryCatalogStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

metadata API calls

fn update_catalog<'life0, 'async_trait>( &'life0 self, request: Request<UpdateCatalogRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<UpdateCatalogResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_recording_schema<'life0, 'async_trait>( &'life0 self, request: Request<GetRecordingSchemaRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetRecordingSchemaResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn register_recording<'life0, 'async_trait>( &'life0 self, request: Request<RegisterRecordingRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<DataframePart>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

TODO(zehiko) support registering more than one recording at a time

fn unregister_recording<'life0, 'async_trait>( &'life0 self, request: Request<UnregisterRecordingRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<UnregisterRecordingResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn unregister_all_recordings<'life0, 'async_trait>( &'life0 self, request: Request<UnregisterAllRecordingsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<UnregisterAllRecordingsResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,