re_protos::v1alpha1::rerun_frontend_v1alpha1::frontend_service_server
Trait FrontendService

pub trait FrontendService:
    Send
    + Sync
    + 'static {
    type ScanPartitionTableStream: Stream<Item = Result<ScanPartitionTableResponse, Status>> + Send + 'static;
    type SearchDatasetStream: Stream<Item = Result<SearchDatasetResponse, Status>> + Send + 'static;
    type QueryDatasetStream: Stream<Item = Result<QueryDatasetResponse, Status>> + Send + 'static;
    type GetChunksStream: Stream<Item = Result<GetChunksResponse, Status>> + Send + 'static;
    type ScanTableStream: Stream<Item = Result<ScanTableResponse, Status>> + Send + 'static;
    type QueryTasksOnCompletionStream: Stream<Item = Result<QueryTasksOnCompletionResponse, Status>> + Send + 'static;

    // Required methods
    fn find_entries<'life0, 'async_trait>(
        &'life0 self,
        request: Request<FindEntriesRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<FindEntriesResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn delete_entry<'life0, 'async_trait>(
        &'life0 self,
        request: Request<DeleteEntryRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<DeleteEntryResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn create_dataset_entry<'life0, 'async_trait>(
        &'life0 self,
        request: Request<CreateDatasetEntryRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CreateDatasetEntryResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn read_dataset_entry<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ReadDatasetEntryRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ReadDatasetEntryResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn read_table_entry<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ReadTableEntryRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ReadTableEntryResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn register_with_dataset<'life0, 'async_trait>(
        &'life0 self,
        request: Request<RegisterWithDatasetRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<RegisterWithDatasetResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn write_chunks<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Streaming<WriteChunksRequest>>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<WriteChunksResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_partition_table_schema<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetPartitionTableSchemaRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetPartitionTableSchemaResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn scan_partition_table<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ScanPartitionTableRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::ScanPartitionTableStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_dataset_schema<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetDatasetSchemaRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetDatasetSchemaResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn create_index<'life0, 'async_trait>(
        &'life0 self,
        request: Request<CreateIndexRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CreateIndexResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn re_index<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ReIndexRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ReIndexResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn search_dataset<'life0, 'async_trait>(
        &'life0 self,
        request: Request<SearchDatasetRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::SearchDatasetStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn query_dataset<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryDatasetRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::QueryDatasetStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_chunks<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetChunksRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::GetChunksStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_table_schema<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetTableSchemaRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetTableSchemaResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn scan_table<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ScanTableRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::ScanTableStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn query_tasks<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryTasksRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QueryTasksResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn fetch_task_output<'life0, 'async_trait>(
        &'life0 self,
        request: Request<FetchTaskOutputRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<FetchTaskOutputResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn query_tasks_on_completion<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryTasksOnCompletionRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::QueryTasksOnCompletionStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

}

Generated trait containing gRPC methods that should be implemented for use with FrontendServiceServer.

Required Associated Types

type ScanPartitionTableStream: Stream<Item = Result<ScanPartitionTableResponse, Status>> + Send + 'static

Server streaming response type for the ScanPartitionTable method.

type SearchDatasetStream: Stream<Item = Result<SearchDatasetResponse, Status>> + Send + 'static

Server streaming response type for the SearchDataset method.

type QueryDatasetStream: Stream<Item = Result<QueryDatasetResponse, Status>> + Send + 'static

Server streaming response type for the QueryDataset method.

type GetChunksStream: Stream<Item = Result<GetChunksResponse, Status>> + Send + 'static

Server streaming response type for the GetChunks method.

type ScanTableStream: Stream<Item = Result<ScanTableResponse, Status>> + Send + 'static

Server streaming response type for the ScanTable method.

type QueryTasksOnCompletionStream: Stream<Item = Result<QueryTasksOnCompletionResponse, Status>> + Send + 'static

Server streaming response type for the QueryTasksOnCompletion method.

Required Methods

fn find_entries<'life0, 'async_trait>( &'life0 self, request: Request<FindEntriesRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<FindEntriesResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn delete_entry<'life0, 'async_trait>( &'life0 self, request: Request<DeleteEntryRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<DeleteEntryResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn create_dataset_entry<'life0, 'async_trait>( &'life0 self, request: Request<CreateDatasetEntryRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<CreateDatasetEntryResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn read_dataset_entry<'life0, 'async_trait>( &'life0 self, request: Request<ReadDatasetEntryRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ReadDatasetEntryResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn read_table_entry<'life0, 'async_trait>( &'life0 self, request: Request<ReadTableEntryRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ReadTableEntryResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn register_with_dataset<'life0, 'async_trait>( &'life0 self, request: Request<RegisterWithDatasetRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<RegisterWithDatasetResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Register new partitions with the Dataset

fn write_chunks<'life0, 'async_trait>( &'life0 self, request: Request<Streaming<WriteChunksRequest>>, ) -> Pin<Box<dyn Future<Output = Result<Response<WriteChunksResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Write chunks to one or more partitions.

The partition ID for each individual chunk is extracted from their metadata (rerun.partition_id).

The destination dataset must be provided in the x-rerun-dataset-id header.

fn get_partition_table_schema<'life0, 'async_trait>( &'life0 self, request: Request<GetPartitionTableSchemaRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetPartitionTableSchemaResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the schema of the partition table (i.e. the dataset manifest) itself, not the underlying dataset.

To inspect the data of the partition table, use ScanPartitionTable.

To retrieve the schema of the underlying dataset, use GetDatasetSchema instead.

fn scan_partition_table<'life0, 'async_trait>( &'life0 self, request: Request<ScanPartitionTableRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::ScanPartitionTableStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Inspect the contents of the partition table (i.e. the dataset manifest).

The returned data will follow the schema specified by GetPartitionTableSchema.

fn get_dataset_schema<'life0, 'async_trait>( &'life0 self, request: Request<GetDatasetSchemaRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetDatasetSchemaResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the schema of the dataset.

This is the union of all the schemas from all the underlying partitions. It will contain all the indexes, entities and components present in the dataset.

fn create_index<'life0, 'async_trait>( &'life0 self, request: Request<CreateIndexRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<CreateIndexResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Creates a custom index for a specific column (vector search, full-text search, etc).

fn re_index<'life0, 'async_trait>( &'life0 self, request: Request<ReIndexRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ReIndexResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Recreate an index with the same configuration but (potentially) new data.

fn search_dataset<'life0, 'async_trait>( &'life0 self, request: Request<SearchDatasetRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::SearchDatasetStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Search a previously created index.

fn query_dataset<'life0, 'async_trait>( &'life0 self, request: Request<QueryDatasetRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::QueryDatasetStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Perform Rerun-native queries on a dataset, returning the matching chunk IDs.

These Rerun-native queries include:

Filtering by specific partition and chunk IDs.

Latest-at, range and dataframe queries.

Arbitrary Lance filters.

To fetch the actual chunks themselves, see GetChunks.

Passing chunk IDs to this method effectively acts as a IF_EXIST filter.

fn get_chunks<'life0, 'async_trait>( &'life0 self, request: Request<GetChunksRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::GetChunksStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Perform Rerun-native queries on a dataset, returning the underlying chunks.

These Rerun-native queries include:

Filtering by specific partition and chunk IDs.

Latest-at, range and dataframe queries.

Arbitrary Lance filters.

To fetch only the actual chunk IDs rather than the chunks themselves, see QueryDataset.

fn get_table_schema<'life0, 'async_trait>( &'life0 self, request: Request<GetTableSchemaRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetTableSchemaResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn scan_table<'life0, 'async_trait>( &'life0 self, request: Request<ScanTableRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::ScanTableStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn query_tasks<'life0, 'async_trait>( &'life0 self, request: Request<QueryTasksRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QueryTasksResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

โ Tasks service โ Query the status of submitted tasks

fn fetch_task_output<'life0, 'async_trait>( &'life0 self, request: Request<FetchTaskOutputRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<FetchTaskOutputResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Fetch the output of a completed task

fn query_tasks_on_completion<'life0, 'async_trait>( &'life0 self, request: Request<QueryTasksOnCompletionRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::QueryTasksOnCompletionStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Query the status of submitted tasks as soon as they are no longer pending