re_protos::v1alpha1::rerun_manifest_registry_v1alpha1::manifest_registry_service_server
Trait ManifestRegistryService

pub trait ManifestRegistryService:
    Send
    + Sync
    + 'static {
    type ListPartitionsStream: Stream<Item = Result<ListPartitionsResponse, Status>> + Send + 'static;
    type QueryDatasetStream: Stream<Item = Result<QueryDatasetResponse, Status>> + Send + 'static;
    type FetchPartitionStream: Stream<Item = Result<FetchPartitionResponse, Status>> + Send + 'static;
    type GetChunksStream: Stream<Item = Result<GetChunksResponse, Status>> + Send + 'static;
    type ListPartitionIndexesStream: Stream<Item = Result<ListPartitionIndexesResponse, Status>> + Send + 'static;
    type SearchDatasetStream: Stream<Item = Result<SearchDatasetResponse, Status>> + Send + 'static;

    // Required methods
    fn register_partitions<'life0, 'async_trait>(
        &'life0 self,
        request: Request<RegisterPartitionsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<RegisterPartitionsResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn unregister_partitions<'life0, 'async_trait>(
        &'life0 self,
        request: Request<UnregisterPartitionsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<UnregisterPartitionsResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn list_partitions<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ListPartitionsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::ListPartitionsStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn create_partition_manifests<'life0, 'async_trait>(
        &'life0 self,
        request: Request<CreatePartitionManifestsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CreatePartitionManifestsResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn query_dataset<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryDatasetRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::QueryDatasetStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn fetch_partition<'life0, 'async_trait>(
        &'life0 self,
        request: Request<FetchPartitionRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::FetchPartitionStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_chunks<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetChunksRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::GetChunksStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn create_partition_indexes<'life0, 'async_trait>(
        &'life0 self,
        request: Request<CreatePartitionIndexesRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CreatePartitionIndexesResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn list_partition_indexes<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ListPartitionIndexesRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::ListPartitionIndexesStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn search_dataset<'life0, 'async_trait>(
        &'life0 self,
        request: Request<SearchDatasetRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::SearchDatasetStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with ManifestRegistryServiceServer.

Required Associated Types

type ListPartitionsStream: Stream<Item = Result<ListPartitionsResponse, Status>> + Send + 'static

Server streaming response type for the ListPartitions method.

type QueryDatasetStream: Stream<Item = Result<QueryDatasetResponse, Status>> + Send + 'static

Server streaming response type for the QueryDataset method.

type FetchPartitionStream: Stream<Item = Result<FetchPartitionResponse, Status>> + Send + 'static

Server streaming response type for the FetchPartition method.

type GetChunksStream: Stream<Item = Result<GetChunksResponse, Status>> + Send + 'static

Server streaming response type for the GetChunks method.

type ListPartitionIndexesStream: Stream<Item = Result<ListPartitionIndexesResponse, Status>> + Send + 'static

Server streaming response type for the ListPartitionIndexes method.

type SearchDatasetStream: Stream<Item = Result<SearchDatasetResponse, Status>> + Send + 'static

Server streaming response type for the SearchDataset method.

Required Methods

fn register_partitions<'life0, 'async_trait>( &'life0 self, request: Request<RegisterPartitionsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<RegisterPartitionsResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Register new partitions with the Dataset

fn unregister_partitions<'life0, 'async_trait>( &'life0 self, request: Request<UnregisterPartitionsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<UnregisterPartitionsResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Unregister partitions from the Dataset

fn list_partitions<'life0, 'async_trait>( &'life0 self, request: Request<ListPartitionsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::ListPartitionsStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

List partitions in the Dataset

fn create_partition_manifests<'life0, 'async_trait>( &'life0 self, request: Request<CreatePartitionManifestsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<CreatePartitionManifestsResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Create manifests for all partitions in the Dataset

fn query_dataset<'life0, 'async_trait>( &'life0 self, request: Request<QueryDatasetRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::QueryDatasetStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Query Dataset and return a dataframe containing relevant chunk IDs

fn fetch_partition<'life0, 'async_trait>( &'life0 self, request: Request<FetchPartitionRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::FetchPartitionStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get all chunks for a single partition

fn get_chunks<'life0, 'async_trait>( &'life0 self, request: Request<GetChunksRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::GetChunksStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get chunks based on a specific query. This call is similar to QueryDataset but instead of returning a dataframe with relevant chunk IDs, it returns the actual chunks.

fn create_partition_indexes<'life0, 'async_trait>( &'life0 self, request: Request<CreatePartitionIndexesRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<CreatePartitionIndexesResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Create index for partitions in the Dataset. Index can be created for all or specific partitions

fn list_partition_indexes<'life0, 'async_trait>( &'life0 self, request: Request<ListPartitionIndexesRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::ListPartitionIndexesStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

List indexes for the Dataset

fn search_dataset<'life0, 'async_trait>( &'life0 self, request: Request<SearchDatasetRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::SearchDatasetStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Do a full text, vector or scalar search. Currently only an Indexed search is supported, user must first call CreatePartitionIndexes for the relevant column. TODO(zehiko) add support for “brute force” search. The response is a RecordBatch with 4 columns:

‘partition_id’ - which partition the data is from

‘timepoint’ - represents the points in time where index query matches. What time points are matched depends on the type of index that is queried. For example: for vector search it might be timepoints where top-K matches are found within each partition in the indexed entry. For inverted index it might be timepoints where the query string is found in the indexed column

instance column - if index column contains a batch of values (for example a list of embeddings), then each instance of the batch is a separate row in the resulting RecordBatch

‘instance_id’ - this is a simple element index in the batch array. For example if indexed column is a list of embeddings [a,b,c] (where each embedding is of same length) then ‘instance_id’ of embedding ‘a’ is 0, ‘instance_id’ of ‘b’ is 1, etc.