re_protos::v1alpha1
Module rerun_redap_tasks_v1alpha1

Modules

tasks_service_client

Generated client implementations.

tasks_service_server

Generated server implementations.

Structs

FetchTaskOutputRequest

FetchTaskOutputRequest is the request message for fetching task output

FetchTaskOutputResponse

/ FetchTaskOutputResponse is the response message for fetching task output

QueryTasksOnCompletionRequest

QueryTasksOnCompletionRequest is the request message for querying tasks status. This is close-to-a-copy of QueryTasksRequest, with the addition of a timeout.

QueryTasksOnCompletionResponse

QueryTaskOnCompletionResponse is the response message for querying tasks status encoded as a record batch. This is a copy of QueryTasksResponse.

QueryTasksRequest

QueryTasksRequest is the request message for querying tasks status

QueryTasksResponse

QueryTasksResponse is the response message for querying tasks status encoded as a record batch

SubmitTasksRequest

SubmitTasksRequest is the request message for submitting tasks

SubmitTasksResponse

SubmitTaskResponse contains, for each submitted task its submission outcome, encoded as a RecordBatch

Task

A task is a unit of work that can be submitted to the system