re_protos::v1alpha1::rerun_redap_tasks_v1alpha1::tasks_service_server
Trait TasksService

pub trait TasksService:
    Send
    + Sync
    + 'static {
    type QueryTasksOnCompletionStream: Stream<Item = Result<QueryTasksOnCompletionResponse, Status>> + Send + 'static;

    // Required methods
    fn submit_tasks<'life0, 'async_trait>(
        &'life0 self,
        request: Request<SubmitTasksRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<SubmitTasksResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn query_tasks<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryTasksRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<QueryTasksResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn fetch_task_output<'life0, 'async_trait>(
        &'life0 self,
        request: Request<FetchTaskOutputRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<FetchTaskOutputResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn query_tasks_on_completion<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryTasksOnCompletionRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::QueryTasksOnCompletionStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with TasksServiceServer.

Required Associated Types

type QueryTasksOnCompletionStream: Stream<Item = Result<QueryTasksOnCompletionResponse, Status>> + Send + 'static

Server streaming response type for the QueryTasksOnCompletion method.

Required Methods

fn submit_tasks<'life0, 'async_trait>( &'life0 self, request: Request<SubmitTasksRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<SubmitTasksResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Submit new tasks

fn query_tasks<'life0, 'async_trait>( &'life0 self, request: Request<QueryTasksRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<QueryTasksResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Query the status of submitted tasks

fn fetch_task_output<'life0, 'async_trait>( &'life0 self, request: Request<FetchTaskOutputRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<FetchTaskOutputResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Fetch the output of a completed task

fn query_tasks_on_completion<'life0, 'async_trait>( &'life0 self, request: Request<QueryTasksOnCompletionRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::QueryTasksOnCompletionStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Query the status of submitted tasks, waiting for their completion.

The method returns a stream of QueryResult. Each item in the stream contains the status of a subset of the tasks, as they complete. The server does not guarantee to immediately send one stream item as soon as a task completes, but may decide to arbitrarily aggregate results into larger batches.