re_protos::v1alpha1
Module rerun_remote_store_v1alpha1

Modules

index_properties

Nested message and enum types in IndexProperties.

index_query_properties

Nested message and enum types in IndexQueryProperties.

storage_node_service_client

Generated client implementations.

storage_node_service_server

Generated server implementations.

Structs

BTreeIndex

TODO(zehiko) add properties as needed

BTreeIndexQuery

TODO(zehiko) add properties as needed

CatalogEntry

CatalogFilter

Collection

A collection in the catalog

CollectionFilter

ColumnProjection

CreateIndexRequest

CreateIndexResponse

CreateManifestsRequest

CreateManifestsResponse

DataframePart

DataframePart is arrow IPC encoded RecordBatch

FetchRecordingRequest

FetchRecordingResponse

TODO(jleibs): Eventually this becomes either query-mediated in some way, but for now it’s useful to be able to just get back the whole RRD somehow.

FindCollectionRequest

FindCollectionResponse

Returns matching collection, if there is exactly one match

FindCollectionsRequest

FindCollectionsResponse

Contains zero or more matching collections

GetChunksRangeRequest

GetChunksRange is a streaming API that allows to fetch chunks within a time range

GetChunksRangeResponse

GetChunksRequest

GetChunksResponse

GetRecordingSchemaRequest

GetRecordingSchemaResponse

IndexColumn

used to define which column we want to index

IndexProperties

IndexQueryProperties

InvertedIndex

InvertedIndexQuery

TODO(zehiko) add properties as needed

ListManifestsRequest

ListManifestsResponse

QueryCatalogRequest

QueryCatalogResponse

QueryManifestLatestAtRelevantChunks

QueryManifestRangeRelevantChunks

QueryManifestRequest

A manifest query will find all the relevant chunk IDs (and optionally a bunch of related metadata) for a given Rerun query (latest-at, range, etc).

QueryManifestResponse

QueryRequest

QueryResponse

ReIndexRequest

ReIndexResponse

RegisterRecordingDescription

RegisterRecordingsRequest

RegisterRecordingsResponse

RemoteStoreError

Application level error - used as details in the google.rpc.Status message

ScanParameters

Generic parameters that will influence the behavior of the Lance scanner.

ScanParametersOrderClause

SearchIndexRequest

SearchIndexResponse

UnregisterAllRecordingsRequest

UnregisterAllRecordingsResponse

UnregisterRecordingRequest

UnregisterRecordingResponse

UpdateCatalogRequest

UpdateCatalogResponse

VectorIndexQuery

VectorIvfPqIndex

Enums

ErrorCode

Error codes for application level errors

RecordingType

VectorDistanceMetric