re_protos::v1alpha1::rerun_remote_store_v1alpha1::storage_node_service_server
Trait StorageNodeService

pub trait StorageNodeService:
    Send
    + Sync
    + 'static {
    type QueryStream: Stream<Item = Result<QueryResponse, Status>> + Send + 'static;
    type FetchRecordingStream: Stream<Item = Result<FetchRecordingResponse, Status>> + Send + 'static;
    type GetChunksStream: Stream<Item = Result<GetChunksResponse, Status>> + Send + 'static;
    type GetChunksRangeStream: Stream<Item = Result<GetChunksRangeResponse, Status>> + Send + 'static;
    type SearchIndexStream: Stream<Item = Result<SearchIndexResponse, Status>> + Send + 'static;
    type ListManifestsStream: Stream<Item = Result<ListManifestsResponse, Status>> + Send + 'static;
    type QueryManifestStream: Stream<Item = Result<QueryManifestResponse, Status>> + Send + 'static;
    type QueryCatalogStream: Stream<Item = Result<QueryCatalogResponse, Status>> + Send + 'static;

    // Required methods
    fn find_collection<'life0, 'async_trait>(
        &'life0 self,
        request: Request<FindCollectionRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<FindCollectionResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn find_collections<'life0, 'async_trait>(
        &'life0 self,
        request: Request<FindCollectionsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<FindCollectionsResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn query<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::QueryStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn fetch_recording<'life0, 'async_trait>(
        &'life0 self,
        request: Request<FetchRecordingRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::FetchRecordingStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn create_index<'life0, 'async_trait>(
        &'life0 self,
        request: Request<CreateIndexRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CreateIndexResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn re_index<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ReIndexRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ReIndexResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_chunks<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetChunksRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::GetChunksStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_chunks_range<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetChunksRangeRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::GetChunksRangeStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn search_index<'life0, 'async_trait>(
        &'life0 self,
        request: Request<SearchIndexRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::SearchIndexStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn create_manifests<'life0, 'async_trait>(
        &'life0 self,
        request: Request<CreateManifestsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CreateManifestsResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn list_manifests<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ListManifestsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::ListManifestsStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn query_manifest<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryManifestRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::QueryManifestStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn query_catalog<'life0, 'async_trait>(
        &'life0 self,
        request: Request<QueryCatalogRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::QueryCatalogStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn update_catalog<'life0, 'async_trait>(
        &'life0 self,
        request: Request<UpdateCatalogRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<UpdateCatalogResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_recording_schema<'life0, 'async_trait>(
        &'life0 self,
        request: Request<GetRecordingSchemaRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetRecordingSchemaResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn register_recordings<'life0, 'async_trait>(
        &'life0 self,
        request: Request<RegisterRecordingsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<RegisterRecordingsResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn unregister_recording<'life0, 'async_trait>(
        &'life0 self,
        request: Request<UnregisterRecordingRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<UnregisterRecordingResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn unregister_all_recordings<'life0, 'async_trait>(
        &'life0 self,
        request: Request<UnregisterAllRecordingsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<UnregisterAllRecordingsResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

}

Generated trait containing gRPC methods that should be implemented for use with StorageNodeServiceServer.

Required Associated Types

type QueryStream: Stream<Item = Result<QueryResponse, Status>> + Send + 'static

Server streaming response type for the Query method.

type FetchRecordingStream: Stream<Item = Result<FetchRecordingResponse, Status>> + Send + 'static

Server streaming response type for the FetchRecording method.

type GetChunksStream: Stream<Item = Result<GetChunksResponse, Status>> + Send + 'static

Server streaming response type for the GetChunks method.

type GetChunksRangeStream: Stream<Item = Result<GetChunksRangeResponse, Status>> + Send + 'static

Server streaming response type for the GetChunksRange method.

type SearchIndexStream: Stream<Item = Result<SearchIndexResponse, Status>> + Send + 'static

Server streaming response type for the SearchIndex method.

type ListManifestsStream: Stream<Item = Result<ListManifestsResponse, Status>> + Send + 'static

Server streaming response type for the ListManifests method.

type QueryManifestStream: Stream<Item = Result<QueryManifestResponse, Status>> + Send + 'static

Server streaming response type for the QueryManifest method.

type QueryCatalogStream: Stream<Item = Result<QueryCatalogResponse, Status>> + Send + 'static

Server streaming response type for the QueryCatalog method.

Required Methods

fn find_collection<'life0, 'async_trait>( &'life0 self, request: Request<FindCollectionRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<FindCollectionResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn find_collections<'life0, 'async_trait>( &'life0 self, request: Request<FindCollectionsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<FindCollectionsResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn query<'life0, 'async_trait>( &'life0 self, request: Request<QueryRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::QueryStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn fetch_recording<'life0, 'async_trait>( &'life0 self, request: Request<FetchRecordingRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::FetchRecordingStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn create_index<'life0, 'async_trait>( &'life0 self, request: Request<CreateIndexRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<CreateIndexResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn re_index<'life0, 'async_trait>( &'life0 self, request: Request<ReIndexRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ReIndexResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_chunks<'life0, 'async_trait>( &'life0 self, request: Request<GetChunksRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::GetChunksStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_chunks_range<'life0, 'async_trait>( &'life0 self, request: Request<GetChunksRangeRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::GetChunksRangeStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn search_index<'life0, 'async_trait>( &'life0 self, request: Request<SearchIndexRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::SearchIndexStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

The response to SearchIndex a RecordBatch with 3 columns:

‘resource_id’ column with the id of the resource

timepoint column with the values representing the points in time where index query matches. What time points are matched depends on the type of index that is queried. For example for vector search it might be timepoints where top-K matches are found within each resource in the indexed entry. For inverted index it might be timepoints where the query string is found in the indexed column

‘data’ column with the data that is returned for the matched timepoints

fn create_manifests<'life0, 'async_trait>( &'life0 self, request: Request<CreateManifestsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<CreateManifestsResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn list_manifests<'life0, 'async_trait>( &'life0 self, request: Request<ListManifestsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::ListManifestsStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn query_manifest<'life0, 'async_trait>( &'life0 self, request: Request<QueryManifestRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::QueryManifestStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn query_catalog<'life0, 'async_trait>( &'life0 self, request: Request<QueryCatalogRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::QueryCatalogStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn update_catalog<'life0, 'async_trait>( &'life0 self, request: Request<UpdateCatalogRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<UpdateCatalogResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn get_recording_schema<'life0, 'async_trait>( &'life0 self, request: Request<GetRecordingSchemaRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<GetRecordingSchemaResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn register_recordings<'life0, 'async_trait>( &'life0 self, request: Request<RegisterRecordingsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<RegisterRecordingsResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn unregister_recording<'life0, 'async_trait>( &'life0 self, request: Request<UnregisterRecordingRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<UnregisterRecordingResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

TODO(cmc): This needs to be batched. Everything needs to be batched, always.

fn unregister_all_recordings<'life0, 'async_trait>( &'life0 self, request: Request<UnregisterAllRecordingsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<UnregisterAllRecordingsResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,