Crate re_protos

This crate contains generated types for the remote store gRPC service API. Generation is done using the re_protos_builder crate.

We want clear separation between ‘internal’ types and gRPC types and don’t want to use gRPC types in the rerun viewer codebase. That’s why we implement all the necessary conversion code (in the form of From and TryFrom traits) in this crate.

Modules

catalog

common

external

frontend

log_msg

manifest_registry

redap_tasks

sdk_comms

sizes 🔒

v1alpha1 🔒

Macros

invalid_field

missing_field

Enums

TypeConversionError