re_protos::v1alpha1
Module rerun_common_v1alpha1

Structs

ApplicationId

ComponentDescriptor

DataframePart

DataframePart is arrow IPC encoded RecordBatch

DatasetHandle

Entry point for all ManifestRegistryService APIs

EntityPath

The unique identifier of an entity, e.g. camera/3/points See <https://www.rerun.io/docs/concepts/entity-path> for more on entity paths.

EntryId

IndexColumnSelector

IndexRange

PartitionId

Unique identifier for a partition. Can be user defined which means it can be of any type. For simplicity we start with a string, but we will probably revisit this.

RerunChunk

RerunChunk is arrow IPC encoded RecordBatch that has rerun-specific semantic constraints and can be directly converted to a Rerun Chunk (re_chunk::Chunk)

ScanParameters

Generic parameters that will influence the behavior of the Lance scanner.

ScanParametersOrderClause

Schema

arrow IPC serialized schema

StoreId

TableId

uniquely identifies a table

TaskId

Unique identifier of a task submitted in the redap tasks subsystem

TimeRange

A time range between start and end time points. Each 64 bit number can represent different time point data depending on the timeline it is associated with. Time range is inclusive for both start and end time points.

Timeline

A recording can have multiple timelines, each is identified by a name, for example log_tick, log_time, etc.

Tuid

Enums

EncoderVersion

supported encoder versions for encoding data See RerunData and RerunChunkData for its usage

IfDuplicateBehavior

Specify how the relevant creation call behaves in case of previously created (duplicate) items

IfMissingBehavior

StoreKind